requirements.txt
alembic==1.12.0
requests==2.25.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21